More Forestry compat
Made more (basically all) crops able to be fermented. Values are the same as for all the other pam's crops. Also maintained fermentation with various fluids. Additionally made melons give fruit juice at a rate of 15mb per slice. The fermentation script is written in a way to make future additions or substractions as easy as adding/removing entries from the input arrays.