🌍 Worldgen: Bedrock  1 => 2 layers, but 1 layer in Nehter
This is related to:
https://github.com/NillerMedDild/Enigmatica2Expert/issues/1899

**Astral Sorcery** rituals not working without clar sky
OR 1 layer of bedrock above ritual.

I make floor and roof with 2 blocks, but replace 1 layer of
[Bedrock] with [Compressed Netherrack] in nether to give players ability to use rituals.

This change necessary to prevent making holes in **Overworld** with
[Mining Laser] or many other block miners, because they remove
**Bedrock Ores** and leave holes.

Now `y=0` level is plain bedrock, and `y=1` level content Ores.

Also, fix [Draconium Ore] spawn in **Deep Dark**