🧩 Bansoukou: fix most erroring missed achievments
Before this change, achievements that had a reward in the form of a modified recipe simply wouldn't load.
If such a "mistaken" achievement had children, the whole branch was missing.

The Bansoukou mod changes the `.json` files inside the `.jar` files, allowing achievements to be changed before they are loaded.
This reduces the number of errors in the log and makes more achievements available.
This change completely removes the error `Parsing error loading built-in advancement` from the log.

Example - fully visible Integrated Dynamics branch https://i.imgur.com/QBLpvR2.png