OTG Dimension Configs
Two dimension configs that OTG  uses when running BBOP-Extended or when a vanilla overworld is made. A void world is made in DIM-3 and Alien Jungle in DIM-200, if the Alien Jungle mod is installed.

OTG Dimensions Config and Alien Jungle

Two dimension configs that OTG uses when running BBOP-Extended or when a vanilla overworld is made, and the Alien Jungle world config for OTG. A void world is made in DIM-3 and Alien Jungle in DIM-200.