fixes to old save loading, void world (#207)
By https://github.com/git-confused

fixes #206, #205, #204

the main thing is that old saves should now load without any save editing

islands are once again spawned by voidislandcontrol (because with the fix to #205, JED can't remember it already spawned the island)

removes void teleport to overworld from skyblock

skyblock instructions are simplified (it is harmless, but not necessary to set the world type)

In a future version, the void preset files can be freely removed from the pack. Any savegame loaded in this version will be updated automatically by otg to remove the void otg dim (and JED automatically replaces it).

I only tested this on my old save, so someone else should test it on theirs.

I don't know the meaning of the differently named files in config/OpenTerrainGenerator, so please correct me if deleting them is wrong.

OTG has been bansoukoued to not delete region data when it forgets a dimension. OTG only needs to go through this save update process once, so if OTG receives an update we need to change the config again to not delete any bases.

Also, hide alien jungle, +otg config version, remove seed